requirements.txt
torch>=1.13.0
transformers>=4.30.0
Pillow>=9.0.0
requests>=2.28.0
numpy>=1.21.0
modelscope>=1.0.0